Refactor the private-messaging code for clarity and robustness.
In particular, subscribing to a new channel can apparently
cause double-deliveries of messages from the main channel.
Solution: add a "message_id" field and deduplication logic.

Handshaking with oneself is allowed (and opens a private channel
between yourself and yourself). This does generate some extra
traffic, which can result in a lot of retries due to PubNub's
rate limiting; so make sure handshaking happens out-of-band on
a timer instead of blocking the "Foo entered the room" announcement
(otherwise Foo's first public message could show up before the
"entered the room" announcement).

Change the UI to refresh the "PM Foo" button every time a new
user named "Foo" completes a handshake. This ensures that if
Foo quits his session without sending an "exit" stage direction,
and then "enters" a second time, you won't end up with either
a nonworking "PM Foo" button or a pair of "PM Foo" buttons.
The downside is that a troll could enter the room with the same
name as an existing user and thereby hijack that user's PM
buttons. Therefore, colorize each PM button in the color associated
with the associated user's id.