Add "scroll lock".
It turns out to be useful to have the history scroll automatically,
but I originally didn't want to enable scrolling, so that you could
leave the chat window at a certain point (e.g. to switch tabs) and
not lose your place the next time someone spoke.

So, add a button that does the right thing: allows the user to
control whether autoscroll is on or off.