Better UI for Waving Hands.
The name "Waving Hands Referee" is now slightly magic to PingChat.
When someone by that name enters the room and handshakes with you,
you'll receive a "Play Waving Hands" button in the upper right corner
of the screen. Clicking this button will send a private message to
the referee asking to join, and simultaneously open up the full
button-menu of Waving Hands commands.

If the referee responds "Sorry" to your request, the button-menu
will automatically close up again.

Also, begin implementing spells. This will be incremental: add
one spell and all its interactions, then the next spell, and so on.