Don't wait to resolve orders; plus many minor bugfixes.
Old behavior: if Bob is slow in giving orders, Alice can issue
many sets of orders serially, and only the last set will "take".
New behavior: Alice's orders are locked in as soon as she has
given an order for the left hand and an order for the right hand.
This means that we can immediately send Alice the private messages
asking her to pick spells and targets, instead of having to wait
for Bob.

Also minor bugfixes:
- Surrendering this turn shouldn't immediately cause a wizard to
be !w.isStillFighting(), so we need a second property `is_surrendering`.
- Fix the wording of a comment.

TODO: what happens if Alice still continues trying to change her
gestures? No-op in the engine; should the referee complain?